ABW-1398 - Composable extensions moved to Extension file.
