+++ app/src/main/java/com/babylon/wallet/android/utils/ContextExtensions.kt
@@ -3,22 +3,13 @@ package com.babylon.wallet.android.utils
 import android.content.Context
 import android.content.ContextWrapper
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
-import androidx.compose.ui.platform.LocalContext
 import androidx.fragment.app.FragmentActivity
-import coil.compose.AsyncImagePainter
-import coil.decode.SvgDecoder
-import coil.request.ImageRequest
-import com.babylon.wallet.android.BuildConfig
-
-private const val IMAGE_RATIO = 9 / 16f
 
 fun Context.findFragmentActivity(): FragmentActivity? {
     var context = this
@@ -42,49 +33,5 @@ inline fun Modifier.throttleClickable(
                 lastClickMs = now
             }
         }
-    }
-}
-
-fun Modifier.applyImageAspectRatio(
-    painter: AsyncImagePainter
-): Modifier {
-    return then(
-        (painter.state as? AsyncImagePainter.State.Success)
-            ?.painter?.intrinsicSize?.let { intrinsicSize ->
-                if (intrinsicSize.width / intrinsicSize.height < IMAGE_RATIO) {
-                    Modifier.aspectRatio(IMAGE_RATIO)
-                } else {
-                    Modifier
-                }
-            } ?: Modifier
-    )
-}
-
-// For some reason Coil library requires this header to be added when using with cloudflare service. Otherwise it fails
-private fun Context.buildImageRequest(imageUrl: String?): ImageRequest {
-    return ImageRequest.Builder(this)
-        .data(imageUrl)
-        .decoderFactory(SvgDecoder.Factory())
-        .addHeader("accept", "text/html")
-        .build()
-}
-
-@Composable
-fun rememberImageUrl(fromUrl: String?, size: ImageSize = ImageSize.SMALL): ImageRequest {
-    val context = LocalContext.current
-    val url = "${BuildConfig.IMAGE_HOST_BASE_URL}/?imageOrigin=$fromUrl&imageSize=${size.toSizeString()}"
-    return remember(url) {
-        context.buildImageRequest(url)
-    }
-}
-
-@Suppress("MagicNumber")
-enum class ImageSize(val size: Int) {
-    SMALL(112),
-    MEDIUM(256),
-    LARGE(512);
-
-    fun toSizeString(): String {
-        return "${size}x$size"
     }
 }
